// src/app/credit-request/models/uIOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DashboardModule } from './dashboardModule';
import { EligibleProductOutputModel } from './eligibleProductOutputModel';
import { ProductRequestOutputModel } from './productRequestOutputModel';

/**
 * Information necessary to construct the application shell
 */
export interface UIOutputModel { 
    /**
     * The customer Id for which the info is applicable
     */
    customerId?: string;
    /**
     * The user Id for which the info is applicable
     */
    userId?: string;
    /**
     * An array that lists products for which the user can apply,  If empty array, the user is not entitled to apply for products or the user has a request in progress for the product
     */
    eligibleProducts?: Array<EligibleProductOutputModel>;
    /**
     * Will generate a lists o requests in progress
     */
    requestsInProgress?: Array<ProductRequestOutputModel>;
    /**
     * List of modules to be displayed on the dashboard of the application
     */
    dashboardModules?: Array<DashboardModule>;
}